Add cookie management and user state handling
Integrated `ngx-cookie-service` for cookie management in the Angular app. Updated `LoginComponent` to set an "Authorization" cookie and navigate upon successful login. Enhanced `AuthService` with a `$user` BehaviorSubject for user state management and added a `setUser` method to store user details in `localStorage`. Introduced a `User` interface to define user structure.